+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * Relax semicolon rule for one line methods, classes and modules
+* New cop ClassMethods checks for uses for class/module names in definitions of class/module methods
 
 ### Bugs fixed